fix(modelka): lab1

+++ modelka/01/main.cpp
@@ -40,7 +40,7 @@ int main() {
       }
     }
 
-    int total_wait = PC1.T_wait + PC2.T_wait;
+    double total_wait = PC1.T_wait + PC2.T_wait;
     P_total_wait += total_wait / 1000.0;
     P_in_sys += total_proccess / 1000.0;
 
@@ -50,8 +50,8 @@ int main() {
 
   cout << "Average waiting time: " << P_total_wait / 100.0 << "s" << endl;
   cout << "Average time of task in sys: " << P_in_sys / 100.0 << "s" << endl;
-  cout << "Probability of downtime PC1: " << P_PC1_downtime / 100 << "%" << endl;
-  cout << "Probability of downtime PC2: " << P_PC2_downtime / 100 << "%" << endl;
+  cout << "Probability of downtime PC1: " << P_PC1_downtime / 100.0 << "%" << endl;
+  cout << "Probability of downtime PC2: " << P_PC2_downtime / 100.0 << "%" << endl;
 
   return 0;
 }
